Move to partial class
